Split the client into modules
This makes it much easier to read the code for each section individually.
