+++ pb4py/__init__.py
@@ -1,5 +1,5 @@
 from pb4py import auth, exceptions
-from pb4py.base import Client
+from pb4py.client import Client
 
 __all__ = [
 	'auth',
